requirements.txt
streamlit
numpy
matplotlib
pillow
opencv-python-headless